+++ app/Http/Controllers/hobbyTagController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Tag;
 use App\Hobby;
+use Illuminate\Support\Facades\Gate;
 
 class hobbyTagController extends Controller
 {
@@ -24,6 +25,11 @@ class hobbyTagController extends Controller
 
     public function attachTag($hobby_id, $tag_id){
         $hobby = Hobby::find($hobby_id);
+
+        if(Gate::denies('connect_hoobyTag', $hobby)){
+            abort(403, 'Nonono this hobby is not yours!!!');
+        }
+
         $tag = Tag::find($tag_id);
         $hobby->tags()->attach($tag_id);
         return back()->with([
@@ -32,6 +38,11 @@ class hobbyTagController extends Controller
     }
     public function detachTag($hobby_id, $tag_id){
         $hobby = Hobby::find($hobby_id); 
+
+        if(Gate::denies('connect_hoobyTag', $hobby)){
+            abort(403, 'Nonono this hobby is not yours!!!');
+        }
+
         $tag = Tag::find($tag_id);
         $hobby->tags()->detach($tag_id);
         return back()->with([
